Refactor: TeamHP를 TeamContext로 이동
