+++ src/contexts/TeamStateContext/TeamStateContext.tsx
@@ -5,9 +5,13 @@ interface TeamStateContextType {
     teamOneUsers: (TeamUser | null)[];
     teamTwoUsers: (TeamUser | null)[];
     updateTeams: (users: TeamUser[]) => void;
-    updateAttackTeam : (attackteam : TeamType) => void;
+    updateAttackTeam: (attackteam: TeamType) => void;
     isTeamsLoaded: boolean;
     attackTeam: TeamType | null;
+    // HP
+    teamOneHealth: number;
+    teamTwoHealth: number;
+    changeTeamHP: (team: TeamType, health: number) => void;
 }
 
 const TeamStateContext = createContext<TeamStateContextType | null>(null);
@@ -16,7 +20,12 @@ export const TeamStateProvider = ({ children }: { children: ReactNode }) => {
     const [teamOneUsers, setTeamOneUsers] = useState<(TeamUser | null)[]>(Array(5).fill(null));
     const [teamTwoUsers, setTeamTwoUsers] = useState<(TeamUser | null)[]>(Array(5).fill(null));
     const [isTeamsLoaded, setIsTeamsLoaded] = useState(false);
-    const [attackTeam, setAttackTeam] = useState<TeamType|null>(null);
+    const [attackTeam, setAttackTeam] = useState<TeamType | null>(null);
+
+    // HP 관리
+    //팀 체력
+    const [teamOneHealth, setTeamOneHealth] = useState(3);
+    const [teamTwoHealth, setTeamTwoHealth] = useState(3);
 
     const updateTeams = useCallback((users: TeamUser[]) => {
         const blueTeamUsers = users.filter(user => user.team === 'BLUE');
@@ -35,9 +44,22 @@ export const TeamStateProvider = ({ children }: { children: ReactNode }) => {
         setIsTeamsLoaded(true);
     }, []);
 
-    const updateAttackTeam = useCallback((attackteam : TeamType ) => {
+    const updateAttackTeam = useCallback((attackteam: TeamType) => {
         setAttackTeam(attackteam);
-    },[])
+    }, [])
+
+    const changeTeamHP = (team: TeamType, health: number) => {
+        switch (team) {
+            case "BLUE":
+                setTeamOneHealth(health);
+                break;
+            case "RED":
+                setTeamTwoHealth(health);
+                break;
+            default:
+                break;
+        }
+    }
 
     return (
         <TeamStateContext.Provider value={{
@@ -46,7 +68,10 @@ export const TeamStateProvider = ({ children }: { children: ReactNode }) => {
             updateTeams,
             isTeamsLoaded,
             attackTeam,
-            updateAttackTeam
+            updateAttackTeam,
+            teamOneHealth,
+            teamTwoHealth,
+            changeTeamHP
         }}>
             {children}
         </TeamStateContext.Provider>
